added the chervon,sort arrow and double arrow to render files. Removed all the extra spaces
