--- client/app/components/ChervonIcon.jsx
@@ -1,32 +1,0 @@
-import React from 'react';
-
-/* eslint-disable max-len */
-
-
-export class ChervonDown extends React.PureComponent {
-   // eslint-disable-next-line class-methods-use-this
-  render() {
-    return <svg width="14px" height="10px" className="table-icon" viewBox="0 0 14 10" version="1.1" xmlns="http://www.w3.org/2000/svg">
-    <g stroke="none" strokeWidth="1" fill="none" fillRule="evenodd">
-        <g id="chervon-down-0872B9" fillRule="nonzero" fill="#0872B9">
-           <path d="M12.3130435,2.12560386 L6.57777778,7.8531401 C6.47987118,7.9510467 6.36392915,8 6.22995169,8 C6.09597424,8 5.98003221,7.9510467 5.8821256,7.8531401 L0.146859903,2.12560386 C0.0489533011,2.02769726 0,1.91046699 0,1.77391304 C0,1.6373591 0.0489533011,1.52012882 0.146859903,1.42222222 L1.42995169,0.146859903 C1.52785829,0.0489533011 1.64380032,0 1.77777778,0 C1.91175523,0 2.02769726,0.0489533011 2.12560386,0.146859903 L6.22995169,4.25120773 L10.3342995,0.146859903 C10.4322061,0.0489533011 10.5481481,0 10.6821256,0 C10.8161031,0 10.9320451,0.0489533011 11.0299517,0.146859903 L12.3130435,1.42222222 C12.4109501,1.52012882 12.4599034,1.6373591 12.4599034,1.77391304 C12.4599034,1.91046699 12.4109501,2.02769726 12.3130435,2.12560386 Z" id="Shape"></path>
-        </g>
-    </g>
-  </svg>;
-  }
-}
-
-export class ChervonUp extends React.PureComponent {
-  // eslint-disable-next-line class-methods-use-this
-  render () {
-    return <svg width="13px" height="10px" className="table-icon" viewBox="0 0 17 11" version="1.1" xmlns="http://www.w3.org/2000/svg">
-    <g stroke="none" strokeWidth="1" fill="none" fillRule="evenodd">
-        <g id="chevron-up-0872B9" fillRule="nonzero" fill="#0872B9">
-          <path d="M12.8948156,4.886695 L6.58343916,-1.38094705 C6.35101059,-1.62485579 6.06077157,-1.74675958 5.71214568,-1.74675958 C5.36351979,-1.74675958 5.07328076,-1.62472091 4.8408522,-1.38094705 L4.10515053,-0.658764044 C3.86631365,-0.421262414 3.74689521,-0.132571309 3.74689521,0.207612766 C3.74689521,0.541356005 3.86627974,0.833486719 4.10515053,1.08373514 L8.80976912,5.75327414 L4.10494709,10.4324238 C3.86617802,10.6699254 3.74675958,10.9586165 3.74675958,11.2988343 C3.74675958,11.6325439 3.86614412,11.9247757 4.10494709,12.174923 L4.84071657,12.8970048 C5.07951955,13.1345065 5.37006373,13.2532404 5.71201005,13.2532404 C6.054092,13.2532404 6.34456837,13.1345065 6.58330353,12.8970048 L12.8948156,6.62936279 C13.1337203,6.37904693 13.2532404,6.08698365 13.2532404,5.75324042 C13.2532743,5.41305634 13.1337203,5.12419663 12.8948156,4.886695 Z" id="Shape" transform="translate(8.500000, 5.753240) scale(-1, -1) rotate(90.000000) translate(-8.500000, -5.753240) "></path>
-        </g>
-    </g>
-  </svg>;
-  }
-}
-
-/* eslint-enable max-len */
