+++ client/app/components/RenderFunctions.jsx
@@ -1,3 +1,5 @@
+/* eslint-disable max-lines */
+
 import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
 
@@ -431,6 +433,85 @@ export class Keyboard extends React.PureComponent {
   }
 }
 
-/* eslint-enable max-len */
+export class ChervonDown extends React.PureComponent {
+   // eslint-disable-next-line class-methods-use-this
+  render() {
+    return <svg width="14px" height="10px" className="table-icon" viewBox="0 0 14 10" version="1.1" xmlns="http://www.w3.org/2000/svg">
+    <g stroke="none" strokeWidth="1" fill="none" fillRule="evenodd">
+        <g fillRule="nonzero" fill="#0872B9">
+           <path d="M12.3130435,2.12560386 L6.57777778,7.8531401 C6.47987118,7.9510467 6.36392915,8 6.22995169,8 C6.09597424,8 5.98003221,7.9510467 5.8821256,7.8531401 L0.146859903,2.12560386 C0.0489533011,2.02769726 0,1.91046699 0,1.77391304 C0,1.6373591 0.0489533011,1.52012882 0.146859903,1.42222222 L1.42995169,0.146859903 C1.52785829,0.0489533011 1.64380032,0 1.77777778,0 C1.91175523,0 2.02769726,0.0489533011 2.12560386,0.146859903 L6.22995169,4.25120773 L10.3342995,0.146859903 C10.4322061,0.0489533011 10.5481481,0 10.6821256,0 C10.8161031,0 10.9320451,0.0489533011 11.0299517,0.146859903 L12.3130435,1.42222222 C12.4109501,1.52012882 12.4599034,1.6373591 12.4599034,1.77391304 C12.4599034,1.91046699 12.4109501,2.02769726 12.3130435,2.12560386 Z" id="Shape"></path>
+        </g>
+    </g>
+  </svg>;
+  }
+}
+
+export class ChervonUp extends React.PureComponent {
+  // eslint-disable-next-line class-methods-use-this
+  render () {
+    return <svg width="13px" height="10px" className="table-icon" viewBox="0 0 17 11" version="1.1" xmlns="http://www.w3.org/2000/svg">
+    <g stroke="none" strokeWidth="1" fill="none" fillRule="evenodd">
+        <g fillRule="nonzero" fill="#0872B9">
+          <path d="M12.8948156,4.886695 L6.58343916,-1.38094705 C6.35101059,-1.62485579 6.06077157,-1.74675958 5.71214568,-1.74675958 C5.36351979,-1.74675958 5.07328076,-1.62472091 4.8408522,-1.38094705 L4.10515053,-0.658764044 C3.86631365,-0.421262414 3.74689521,-0.132571309 3.74689521,0.207612766 C3.74689521,0.541356005 3.86627974,0.833486719 4.10515053,1.08373514 L8.80976912,5.75327414 L4.10494709,10.4324238 C3.86617802,10.6699254 3.74675958,10.9586165 3.74675958,11.2988343 C3.74675958,11.6325439 3.86614412,11.9247757 4.10494709,12.174923 L4.84071657,12.8970048 C5.07951955,13.1345065 5.37006373,13.2532404 5.71201005,13.2532404 C6.054092,13.2532404 6.34456837,13.1345065 6.58330353,12.8970048 L12.8948156,6.62936279 C13.1337203,6.37904693 13.2532404,6.08698365 13.2532404,5.75324042 C13.2532743,5.41305634 13.1337203,5.12419663 12.8948156,4.886695 Z" id="Shape" transform="translate(8.500000, 5.753240) scale(-1, -1) rotate(90.000000) translate(-8.500000, -5.753240) "></path>
+        </g>
+    </g>
+  </svg>;
+  }
+}
+
+export class SortArrowUp extends React.PureComponent {
+  // eslint-disable-next-line class-methods-use-this
+  render () {
+    return <svg width="13px" height="10px" className="cf-sort-arrowup table-icon" viewBox="0 0 18 10" version="1.1" xmlns="http://www.w3.org/2000/svg">
+    <g stroke="none" strokeWidth="1" fill="none" fillRule="evenodd">
+        <g fillRule="nonzero" fill="#323A45">
+            <g transform="translate(2.000000, 1.000000)">
+              <path d="M14.2222222,0.888888889 C14.2222222,1.12962963 14.1342593,1.33796296 13.9583333,1.51388889 L7.73611111,7.73611111 C7.56018519,7.91203704 7.35185185,8 7.11111111,8 C6.87037037,8 6.66203704,7.91203704 6.48611111,7.73611111 L0.263888889,1.51388889 C0.087962963,1.33796296 0,1.12962963 0,0.888888889 C0,0.648148148 0.087962963,0.439814815 0.263888889,0.263888889 C0.439814815,0.087962963 0.648148148,0 0.888888889,0 L13.3333333,0 C13.5740741,0 13.7824074,0.087962963 13.9583333,0.263888889 C14.1342593,0.439814815 14.2222222,0.648148148 14.2222222,0.888888889 Z" id="Shape"></path>
+            </g>
+        </g>
+    </g>
+   </svg>;
+  }
+}
+
+export class SortArrowDown extends React.PureComponent {
+  // eslint-disable-next-line class-methods-use-this
+  render () {
+    return <svg width="13px" height="10px" className="cf-sort-arrowdown table-icon " viewBox="0 0 16 9" version="1.1" xmlns="http://www.w3.org/2000/svg">
+    <g stroke="none" strokeWidth="1" fill="none" fillRule="evenodd">
+        <g fillRule="nonzero" fill="#323A45">
+            <g transform="translate(8.500000, 5.000000) scale(1, -1) translate(-8.500000, -5.000000) translate(1.000000, 1.000000)">
+              <path d="M14.2222222,0.888888889 C14.2222222,1.12962963 14.1342593,1.33796296 13.9583333,1.51388889 L7.73611111,7.73611111 C7.56018519,7.91203704 7.35185185,8 7.11111111,8 C6.87037037,8 6.66203704,7.91203704 6.48611111,7.73611111 L0.263888889,1.51388889 C0.087962963,1.33796296 0,1.12962963 0,0.888888889 C0,0.648148148 0.087962963,0.439814815 0.263888889,0.263888889 C0.439814815,0.087962963 0.648148148,0 0.888888889,0 L13.3333333,0 C13.5740741,0 13.7824074,0.087962963 13.9583333,0.263888889 C14.1342593,0.439814815 14.2222222,0.648148148 14.2222222,0.888888889 Z" id="Shape"></path>
+            </g>
+        </g>
+    </g>
+   </svg>;
+  }
+}
+
+export class DoubleArrow extends React.PureComponent {
+// eslint-disable-next-line class-methods-use-this
+  render () {
+    return <svg width="13px" height="16px" className="table-icon" viewBox="0 0 13 16" version="1.1" xmlns="http://www.w3.org/2000/svg">
+    <g stroke="none" strokeWidth="1" fill="none" fillRule="evenodd">
+        <g fillRule="nonzero" fill="#323A45">
+            <g transform="translate(2.000000, 1.000000)">
+                <g transform="translate(5.000000, 2.666667) scale(-1, 1) rotate(-180.000000) translate(-5.000000, -2.666667) ">
+                    <path d="M9.48148147,0.592592593 C9.48148147,0.75308642 9.42283953,0.891975307 9.30555553,1.00925926 L5.15740741,5.15740741 C5.04012346,5.27469136 4.90123457,5.33333333 4.74074074,5.33333333 C4.58024691,5.33333333 4.44135803,5.27469136 4.32407407,5.15740741 L0.175925926,1.00925926 C0.0586419753,0.891975307 0,0.75308642 0,0.592592593 C0,0.432098765 0.0586419753,0.293209877 0.175925926,0.175925926 C0.293209877,0.0586419753 0.432098765,0 0.592592593,0 L8.88888887,0 C9.04938273,0 9.1882716,0.0586419753 9.30555553,0.175925926 C9.42283953,0.293209877 9.48148147,0.432098765 9.48148147,0.592592593 Z" id="Shape"></path>
+                </g>
+            </g>
+            <g transform="translate(7.000000, 11.666667) scale(1, -1) translate(-7.000000, -11.666667) translate(2.000000, 9.000000)">
+                <g transform="translate(5.000000, 2.666667) scale(-1, 1) rotate(-180.000000) translate(-5.000000, -2.666667) ">
+                    <path d="M9.48148147,0.592592593 C9.48148147,0.75308642 9.42283953,0.891975307 9.30555553,1.00925926 L5.15740741,5.15740741 C5.04012346,5.27469136 4.90123457,5.33333333 4.74074074,5.33333333 C4.58024691,5.33333333 4.44135803,5.27469136 4.32407407,5.15740741 L0.175925926,1.00925926 C0.0586419753,0.891975307 0,0.75308642 0,0.592592593 C0,0.432098765 0.0586419753,0.293209877 0.175925926,0.175925926 C0.293209877,0.0586419753 0.432098765,0 0.592592593,0 L8.88888887,0 C9.04938273,0 9.1882716,0.0586419753 9.30555553,0.175925926 C9.42283953,0.293209877 9.48148147,0.432098765 9.48148147,0.592592593 Z" id="Shape"></path>
+                </g>
+            </g>
+        </g>
+    </g>
+ </svg>;
+  }
+}
+
+ /* eslint-enable max-len */
+/* eslint-enable max-lines */
